rename move packets to have "packet" at the end
